# docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Persistance service",
        "version": "0.0.1",
        "description": "Persistance service presenting a REST API to perform authenticated operations on a database. Authorization is granted through the provision of an API key.",
        "contact": {
            "name": "Claire Gouarne",
            "url": "https://github.com/ClrGe"
        }
    },
    "servers": [
        {
            "url": "https://dev.jmg-conseil.eu",
            "description": "Development server"
        }
    ],
    "paths": {
        "/api/docs": {
            "get": {
                "tags": [
                    "DOCS"
                ],
                "description": "OpenAPI 3.0 documentation",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/status": {
            "get": {
                "tags": [
                    "SUPERVISION"
                ],
                "description": "Display service status",
                "parameters": [
                    {
                        "name": "api_key",
                        "in": "header",
                        "description": "Api key provided for authentication",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/stop": {
            "get": {
                "tags": [
                    "SUPERVISION"
                ],
                "parameters": [
                    {
                        "name": "api_key",
                        "in": "header",
                        "description": "Api key provided for authentication",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "description": "Stop service",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/find/all": {
            "get": {
                "tags": [
                    "GET"
                ],
                "description": "List all documents in specified collection of specified database. If not provided, returns default collection.",
                "parameters": [
                    {
                        "name": "api_key",
                        "in": "header",
                        "description": "Api key provided for authentication",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "database",
                        "in": "query",
                        "description": "Name of the database",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "collection",
                        "in": "query",
                        "description": "Name of the collection",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success: return collection",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "POST"
                ],
                "description": "List all documents in specified collection of specified database. If not provided, returns default collection.",
                "parameters": [
                    {
                        "name": "api_key",
                        "in": "header",
                        "description": "Api key provided for authentication",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "collection": {
                                        "type": "string"
                                    },
                                    "database": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success: return collection",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/find": {
            "get": {
                "tags": [
                    "GET"
                ],
                "description": "Display first matching document in specified collection of specified database. If not provided, uses default collection.",
                "parameters": [
                    {
                        "name": "api_key",
                        "in": "header",
                        "description": "Api key provided for authentication",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "database",
                        "in": "query",
                        "description": "Name of the database",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "collection",
                        "in": "query",
                        "description": "Name of the collection",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "POST"
                ],
                "description": "Display first matching document in specified collection of specified database. If not provided, uses default collection.",
                "parameters": [
                    {
                        "name": "api_key",
                        "in": "header",
                        "description": "Api key provided for authentication",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "collection": {
                                        "type": "string"
                                    },
                                    "database": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/find/{id}": {
            "get": {
                "tags": [
                    "GET"
                ],
                "description": "Returns document with id matching query in specified collection of specified database. If not provided, uses default collection.",
                "parameters": [
                    {
                        "name": "api_key",
                        "in": "header",
                        "description": "Api key provided for authentication",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identifier for the target document",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "database",
                        "in": "query",
                        "description": "Name of the database",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "collection",
                        "in": "query",
                        "description": "Name of the collection",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "POST"
                ],
                "description": "Returns document with id matching query in specified collection of specified database. If not provided, uses default collection.",
                "parameters": [
                    {
                        "name": "api_key",
                        "in": "header",
                        "description": "Api key provided for authentication",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identifier for the target document",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "collection": {
                                        "type": "string"
                                    },
                                    "database": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    }
}